replaced input archive by seperate zipped files, included instructions to reproduce the example with other DOOP input
